// packages/dashboard/src/http/types/GetTemplate.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type GetTemplatePath = {
    id: string;
};

export type GetTemplateStatus200 = {
    id: string;
    name: string;
    description: string;
    title: string;
    content: {
        blocks: ({
            type: "paragraph";
            text: string;
            breakBefore?: boolean;
        } | {
            type: "signature-line";
            label?: string;
            breakBefore?: boolean;
        })[];
    };
};

export type GetTemplateOptions = {
    body?: never;
    path: GetTemplatePath;
    query?: never;
    headers?: never;
};

export type GetTemplateResponses = {
    "200": GetTemplateStatus200;
};

/**
 * @description Union of all possible responses
*/
export type GetTemplateResponse = GetTemplateStatus200;